macvim: Patch and enable document icon generation
MacVim's icon generation tool has been failing to compile on OS X 10.5
and caused the brew to fail. This change applies a patch to fix the
problem in makeicns and re-enables icon generation as part of the
standard brew installation.

Upstream, the related issue of downloading the Envy font has been fixed
and made optional. If the download URL changes again in the future, the
Monaco font is used instead.

See http://code.google.com/p/macvim/issues/detail?id=299 for the
upstream ticket.